docker/irods_client/requirements.txt
elabjournal==0.0.19
python-irodsclient==1.1.6
Pillow==9.4.0
pyinstaller==5.8.0
setproctitle==1.3.2
watchdog==2.2.1
